feat: temp commit

+++ chat-server/src/main/java/com/timberliu/chat/server/bean/dto/chat/RecordDTO.java
@@ -10,7 +10,7 @@ import lombok.experimental.Accessors;
 @Data
 public class RecordDTO {
 
-	private Integer msgId;
+	private Long msgId;
 
 	/*
 	 * 文本: 0
